requirements.txt
black
fastapi
openai
pillow
pyTelegramBotAPI
replicate
toml
uvicorn